+++ tests/unit/test_db_models.py
@@ -28,3 +28,28 @@ class TestDbModels:
         flow_doc = FlowDoc(**data)
         assert flow_doc
         assert flow_doc.flow.instructions == flow_dict["instructions"]
+
+    def test_flow_vlan_range(self) -> None:
+        """Test a flow with a vlan range match."""
+        flow_dict = {
+            "match": {
+                "in_port": 9,
+                "dl_vlan": "24/4088",
+                "dl_type": 2048,
+                "nw_proto": 6,
+            },
+            "instructions": [
+                {
+                    "instruction_type": "apply_actions",
+                    "actions": [{"action_type": "output", "port": 4}],
+                },
+            ],
+        }
+        data = {
+            "switch": "00:00:00:00:00:00:00:02",
+            "flow_id": "1",
+            "id": "0",
+            "flow": flow_dict,
+        }
+        flow_doc = FlowDoc(**data)
+        assert flow_doc